allow a function's tensor argument as the output storage
A tensor input argument for a function can be used as the output using
iree.abi.output. In this way, the output can be allocated outside and
a copy can be avoided.

A key change from the compiler side is to be able to inline the wrapped
function for the ABI so that all the computation ops and iree.hal.import
and iree.hal.export ops are in the same scope.

When there is a wrapped function, it is not supported yet since it involves
many other factors to consider.